+++ src/__tests__/plugins-test.js
@@ -94,3 +94,54 @@ test("config extending another config that invokes a plugin with a relative path
 
   t.plan(planned)
 })
+
+test("plugin using exposed rules via stylelint.rules", t => {
+  let planned = 0
+
+  const cssWithDirectiveLower = "/** @@check-color-hex-case */ a { color: #eee; }"
+  const cssWithDirectiveUpper = "/** @@check-color-hex-case */ a { color: #EEE; }"
+  const cssWithoutDirectiveLower = "a { color: #eee; }"
+  const cssWithoutDirectiveUpper = "a { color: #EEE; }"
+  const config = expectation => ({
+    config: {
+      plugins: [path.join(__dirname, "fixtures/plugin-conditionally-check-color-hex-case")],
+      rules: {
+        "conditionally-check-color-hex-case": expectation,
+      },
+    },
+  })
+
+  postcss().use(stylelint(config("upper"))).process(cssWithDirectiveLower).then(result => {
+    t.equal(result.warnings().length, 1)
+    t.equal(result.warnings()[0].text, `Expected "#eee" to be "#EEE" (color-hex-case)`)
+  }).catch(err => console.log(err.stack))
+  planned += 2
+
+  postcss().use(stylelint(config("upper"))).process(cssWithDirectiveUpper).then(result => {
+    t.equal(result.warnings().length, 0)
+  }).catch(err => console.log(err.stack))
+  planned += 1
+
+  postcss().use(stylelint(config("lower"))).process(cssWithDirectiveUpper).then(result => {
+    t.equal(result.warnings().length, 1)
+    t.equal(result.warnings()[0].text, `Expected "#EEE" to be "#eee" (color-hex-case)`)
+  }).catch(err => console.log(err.stack))
+  planned += 2
+
+  postcss().use(stylelint(config("lower"))).process(cssWithDirectiveLower).then(result => {
+    t.equal(result.warnings().length, 0)
+  }).catch(err => console.log(err.stack))
+  planned += 1
+
+  postcss().use(stylelint(config("upper"))).process(cssWithoutDirectiveLower).then(result => {
+    t.equal(result.warnings().length, 0)
+  }).catch(err => console.log(err.stack))
+  planned += 1
+
+  postcss().use(stylelint(config("lower"))).process(cssWithoutDirectiveUpper).then(result => {
+    t.equal(result.warnings().length, 0)
+  }).catch(err => console.log(err.stack))
+  planned += 1
+
+  t.plan(planned)
+})