+++ CHANGELOG.md
@@ -1,17 +1,33 @@
 # Head
 
-- Added: `font-family-name-quotes` rule.
-- Fixed: bug in `rule-properties-order` empty line detection when the two newlines were separated
+* Added: `font-family-name-quotes` rule.
+* Added: better handling of vendor prefixes in `property-unit-blacklist` and `property-unit-whitelist`, e.g. if you enter `animation` it now also checks `-webkit-animation`.
+* Added: support for regular expression property identification in `property-blacklist`, `property-unit-blacklist`, `property-unit-whitelist`, `property-value-blacklist`, and `property-whitelist`.
+* Added: brought `stylelint-rule-tester` into this repo, and exposed it at `stylelint.utils.ruleTester`.
+* Added: `codeFilename` option to Node API.
+* Added: `font-weight-notation` rule.
+* Added: `media-feature-no-missing-punctuation` rule.
+* Added: `no-duplicate-selectors` rule.
+* Added: `no-unknown-animations` rule.
+* Added: `property-value-blacklist` rule.
+* Added: `property-value-whitelist` rule.
+* Added: `time-no-imperceptible` rule.
+* Added: `ignore: "descendant"` and `ignore: "compounded"` options for `selector-no-type`.
+* Added: support for using names of modules for the CLI's `--config` argument, not just paths.
+* Fixed: `ignoreFiles` handling of absolute paths.
+* Fixed: `ignoreFiles` uses the `configBasedir` option to interpret relative paths.
+* Fixed: bug in `rule-properties-order` empty line detection when the two newlines were separated
   by some other whitespace.
-- Fixed: option parsing bug that caused problems when using the `"alphabetical"` primary option
+* Fixed: option parsing bug that caused problems when using the `"alphabetical"` primary option
   with `rule-properties-order`.
+* Fixed: regard an empty string as a valid CSS code
 
 # 4.2.0
 
-- Added: support for custom messages with a `message` secondary property on any rule.
-- Fixed: CLI always ignores contents of `node_modules` and `bower_components` directories.
-- Fixed: bug preventing CLI from understanding absolute paths in `--config` argument.
-- Fixed: bug causing `indentation` to stumble over declarations with semicolons on their own lines.
+* Added: support for custom messages with a `message` secondary property on any rule.
+* Fixed: CLI always ignores contents of `node_modules` and `bower_components` directories.
+* Fixed: bug preventing CLI from understanding absolute paths in `--config` argument.
+* Fixed: bug causing `indentation` to stumble over declarations with semicolons on their own lines.
 
 # 4.1.0